update configuration API with new tagging infrastructure and begin removing experimental multithreading functionality
If you have a use case where the input is large enough for multithreading to be useful,
then you should probably be using ReverseDiff instead of ForwardDiff. The exception to
this is jacobian calculation, but ForwardDiff.jacobian never supported multithreading
anyway. Once Julia's multithreading model matures, we can consider implementing
multithreading for ForwardDiff.jacobian, but we probably should never again support
it for gradients.