"Fix" perturbation confusion, ushering in a new age of glorious dual numbers and prosperity (#213)
* add method for making single seed and improve performance of multi-seed constructor

* refactor Dual to include a tag type parameter

This enables us to throw the proper error(s) in case of perturbation confusion

* unroll config type definitions for readability

* fix binary Dual ambiguities

* fix predicate definition order

* update configuration API with new tagging infrastructure and begin removing experimental multithreading functionality

If you have a use case where the input is large enough for multithreading to be useful,
then you should probably be using ReverseDiff instead of ForwardDiff. The exception to
this is jacobian calculation, but ForwardDiff.jacobian never supported multithreading
anyway. Once Julia's multithreading model matures, we can consider implementing
multithreading for ForwardDiff.jacobian, but we probably should never again support
it for gradients.

* get all tests passing except for deprecation layer + SIMD tests

* drop Julia v0.5

* remove non-v0.6 version compat code and replace test_approx_duals with a simpler predicate (proposed by @KristofferC)

* use where syntax instead of to-be-deprecated prefix typevar syntax

* implement ForwardDiff v0.4.x --> v0.5.x deprecation layer

* update benchmarks

* change T to V in Partials code to avoid confusion with tag parameter

* update docs