Add flexibility in dispatch for `iszero_tuple`
IntervalArithmetic.jl may abandon support for `==` among intervals
(https://github.com/JuliaIntervals/IntervalArithmetic.jl/pull/571).
To support specialization for specific Number subtypes, this makes
`iszero_tuple` into a "trait"-dispatched function, first unwrapping
all the way down to the elementary numeric type and then jointly
dispatching on that type and the actual tuple. This makes it possible
to create an extension in IntervalArithmetic that specializes the
implementation to use the new comparison operator `≛`.
The use of recursive unwrapping enables support for higher-order
derivatives.