Setup GPU testing infrastructure
Noted in https://github.com/JuliaDiff/ForwardDiff.jl/issues/759#issuecomment-3167635946, GPU is completely untested in ForwardDiff.jl, so this sets up the buildkite pipeline. I setup the backend and all, and just took a few tests from https://github.com/JuliaDiff/ForwardDiff.jl/pull/760 to seed it. The point of this isn't really to be a comprehensive set of GPU tests but rather to update this repo to have the standard tools the other repos have so GPU doesn't regress again/more.